# Return fetching AMO result via AM