On networks that do not support atomic operations (AMO) natively the Chapel runtime implements them using active messages (AM). For fetching AMOs (FAMO) the AM handler must return the result of the operation and also set a "done" flag indicating that the result is valid. Previously, the handler did this by doing a PUT of the result back to the initiator locale, performing whatever actions are necessary based on the MCM mode to ensure the PUT is visible, then doing a non-blocking PUT of the done flag. If the MCM mode is delivery-complete the PUT of the result will require one round-trip network latency before the done is PUT. With this change the handler sends a non-blocking AM back to the initiator when possible, which sets the result and the done flag, avoiding the network latency and eliminating one network transfer.

This PR resolves cray/chapel-private#2821 and closes cray/chapel-private#3040.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>