@ronawho 2022-02-14
Some performance results on 16-nodes-c5n-efa for ra-atomics with `xor` switched to `fetchXor` and remote task spawning (500K `coforall loc in Locales do on loc {}`)

```
start_test --perflabel ml- test/release/examples/benchmarks/hpcc/ra-atomics.chpl test/parallel/taskCompare/elliot/empty-chpl-remote-taskspawn.chpl
```
| benchmark           | before | now   |
| ------------------- | -----: | ----: |
| fetching ra-atomics |  28.4s |  2.8s |
| remote task spawn   | 327.1s | 92.7s |

So nice improvement for remote-tasks-spawning due to using inject for non-blocking AMs, and nice speedup for fetching atomics due to switching from PUT+NB-PUT to NB-AM. I think it's worth capturing these perf results in the PR description.